chore(compound): create separate subpackage for compound
refactor(compound): extract field to separate module
